benchmark/03-langgraph-supervisor/requirements.txt
langgraph>=0.2.0,<0.3.0
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
